Ummm why was that in seed.go...
